requirements.txt
cx-freeze==5.0.1
flask==0.12.2
fuzzywuzzy==0.15.1
psutil==5.3.0
pywebview[windows,cocoa]==1.8
requests==2.18.4
xmltodict==0.11.0
